refactor(hive): split the 2.7k-line cli.rs into concern modules
cli.rs becomes cli/mod.rs (HiveCommand enum + dispatch_command + shared
helpers) plus three submodules of command handlers:

- cli/effectiveness.rs  telemetry + clusters (effectiveness, dead-weight, peers, resolutions, convergence, clusters)
- cli/share.rs          share / install / shared-artifact commands
- cli/onboarding.rs     consent, review, explore, experts, welcome

First rewrote in-module `super::` paths to absolute `crate::hive::` (identical
semantics one level up) so handlers survive the move a level deeper; shared
helpers promoted to pub(crate). No public API change beyond re-exporting
share_artifact_from_path (its one external caller). Full suite + clippy
-D warnings green.
